Final system, only needs tiny fixes now.

+++ dmuBlogger/dmuBlogger/Controllers/DevelopersController.cs
@@ -14,12 +14,38 @@ namespace dmuBlogger.Controllers
     public class DevelopersController : Controller
     {
         private DeveloperContext db = new DeveloperContext();
+        private BlacklistContext dbBlacklist = new BlacklistContext();
 
         // GET: Developers
-        public ActionResult Index()
+        public ActionResult Index(string sortOrder, string searchString)
         {
-            return View(db.Developers.ToList());
+            ViewBag.NameSortParm = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewBag.DateSortParm = sortOrder == "Date" ? "date_desc" : "Date";
+            var developers = from s in db.Developers select s;
+
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                developers = developers.Where(s => s.Name.Contains(searchString));
+            }
+
+            switch (sortOrder)
+            {
+                case "DeveloperId":
+                    developers = developers.OrderByDescending(s => s.DeveloperId);
+                    break;
+                case "Name":
+                    developers = developers.OrderBy(s => s.Name);
+                    break;
+                case "Description":
+                    developers = developers.OrderByDescending(s => s.Description);
+                    break;
+                default:
+                    developers = developers.OrderBy(s => s.Name);
+                    break;
+            }
+            return View(developers.ToList());
         }
+        //return View(db.Developers.ToList());
 
         // GET: Developers/Details/5
         public ActionResult Details(int? id)
@@ -49,14 +75,29 @@ namespace dmuBlogger.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "DeveloperId,Name,Description")] Developer developer)
         {
-            if (ModelState.IsValid)
+            bool Banned = false;
+            foreach (Blacklist blacklist in dbBlacklist.Blacklists.ToList())
+            {
+                if (blacklist.BlacklistIP == Request.UserHostAddress)
+                {
+                    Banned = true;
+                }
+            }
+
+            if (ModelState.IsValid && Banned == false)
             {
                 db.Developers.Add(developer);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-
-            return View(developer);
+            if (Banned == true)
+            {
+                return RedirectToAction("../Blacklists/UserBlacklist/");
+            }
+            else
+            {
+                return View(developer);
+            }
         }
 
         // GET: Developers/Edit/5
@@ -81,13 +122,29 @@ namespace dmuBlogger.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "DeveloperId,Name,Description")] Developer developer)
         {
-            if (ModelState.IsValid)
+            bool Banned = false;
+            foreach (Blacklist blacklist in dbBlacklist.Blacklists.ToList())
+            {
+                if (blacklist.BlacklistIP == Request.UserHostAddress)
+                {
+                    Banned = true;
+                }
+            }
+
+            if (ModelState.IsValid && Banned == false)
             {
                 db.Entry(developer).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            return View(developer);
+            if (Banned == true)
+            {
+                return RedirectToAction("../Blacklists/UserBlacklist/");
+            }
+            else
+            {
+                return View(developer);
+            }
         }
 
         // GET: Developers/Delete/5
@@ -110,10 +167,22 @@ namespace dmuBlogger.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Developer developer = db.Developers.Find(id);
-            db.Developers.Remove(developer);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            bool Banned = false;
+            foreach (Blacklist blacklist in dbBlacklist.Blacklists.ToList())
+            {
+                if (blacklist.BlacklistIP == Request.UserHostAddress)
+                {
+                    Banned = true;
+                }
+            }
+            if (Banned == false)
+            {
+                Developer developer = db.Developers.Find(id);
+                db.Developers.Remove(developer);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else { return RedirectToAction("../Blacklists/UserBlacklist/"); }
         }
 
         protected override void Dispose(bool disposing)
